GAT_BiGRU/requirements.txt
tensorflow==2.15.0
pytorch==2.3.1
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
scipy==1.13.1